Added VirtualAddressSpace::freeStack(void*) and VirtualAddressSpace::allocateStack() and implemented those properly for x86 and x64. Only stubs for these functions are available for MIPS. Fixed a bug in the Semaphore code caused by some signedness issues. Removed the VirtualAddressSpace::get/setCurrentAddressSpace() functions and removed all references to these functions. The Thread class now allocates its stacks through the VirtualAddressSpace. All application processors are now brought up (just calling a function from the machine specific interface is missing) on x86 and x64. Fixed the thread/process bug on x64: SS/RSP are pushed/popped independent of the cpl.

+++ src/system/kernel/core/process/Semaphore.cc
@@ -34,7 +34,7 @@ void Semaphore::acquire(size_t n)
 bool Semaphore::tryAcquire(size_t n)
 {
   ssize_t value = m_Counter;
-  if ((value - n) < 0)return false;
+  if ((value - static_cast<ssize_t>(n)) < 0)return false;
 
   if (m_Counter.compareAndSwap(value, value - n))
   {
